GUACAMOLE-337: Move SSL socket into libguac.
