+++ src/libguac/guacamole/socket-ssl.h
